third_party/blink/web_tests/webmidi/requestmidiaccess-options.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../http/tests/resources/permissions-helper.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/media/midi/midi_service.mojom.js"></script>
<script src="../http/tests/webmidi/mock-midiservice.js"></script>
</head>
<body>
<script>
const options = [
    {},
    { sysex: false },
    { sysex: true }
];

const permissions = [
    { midi: 'denied', sysex: 'denied' },
    { midi: 'granted', sysex: 'denied' },
    { midi: 'granted', sysex: 'granted' }
];

promise_test(async _ => {
  for (const permission of permissions) {
    for (const option of options) {
      await PermissionsHelper.setPermission('midi', permission.midi);
      await PermissionsHelper.setPermission('midi-sysex', permission.sysex);

      try {
        const access = await navigator.requestMIDIAccess(option);
        assert_equals(permission.midi, 'granted');
        if (option.sysex) {
          assert_equals(permission.sysex, 'granted');
        }
      } catch (e) {
        if (option.sysex) {
          assert_equals(permission.sysex, 'denied');
        } else {
          assert_equals(permission.midi, 'denied');
        }
      }
    }
  }
}, 'requestMIDIAccess with various options');

</script>
</body>
</html>
